requirements.txt
ccxt>=3.0.0
PyYAML>=6.0
requests>=2.31
numpy>=1.24
pandas>=2.0
